# packages/obsidian-plugin/docs/openapi.yaml
...
components:
  schemas:
    Error:
      properties:
        errorCode:
          description: |
            A 5-digit error code uniquely identifying this particular type of error.
          example: 40149
          type: "number"
        message:
          description: "Message describing the error."
          example: "A brief description of the error."
          type: "string"
      type: "object"
...
    SearchParameters:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: A search phrase for semantic search
          minLength: 1
        filter:
          type: object
          properties:
            folders:
              type: array
              items:
                type: string
              description: 'An array of folder names to include. For example, ["Public", "Work"]'
            excludeFolders:
              type: array
              items:
                type: string
              description: 'An array of folder names to exclude. For example, ["Private", "Archive"]'
            limit:
              type: number
              minimum: 1
              description: The maximum number of results to return
    SearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            type: object
            required:
              - path
              - text
              - score
              - breadcrumbs
            properties:
              path:
                type: string
              text:
                type: string
              score:
                type: number
              breadcrumbs:
                type: string
  securitySchemes:
    apiKeyAuth:
      description: |
        Find your API Key in your Obsidian settings
        in the "Local REST API" section under "Plugins".
      scheme: "bearer"
      type: "http"
info:
  description: |
    You can use this interface for trying out your Local REST API in Obsidian.

    Before trying the below tools, you will want to make sure you press the "Authorize" button below and provide the API Key you are shown when you open the "Local REST API" section of your Obsidian settings.  All requests to the API require a valid API Key; so you won't get very far without doing that.

    When using this tool you may see browser security warnings due to your browser not trusting the self-signed certificate the plugin will generate on its first run.  If you do, you can make those errors disappear by adding the certificate as a "Trusted Certificate" in your browser or operating system's settings.
  title: "Local REST API for Obsidian"
  version: "1.0"
openapi: "3.0.2"
paths:
...
  /search/smart:
    post:
      summary: "Perform a smart search"
      description: "Handles semantic search requests."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchParameters"
      responses:
        "200":
          description: "Successful search"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service unavailable"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
security:
  - apiKeyAuth: []
servers:
  - description: "HTTPS (Secure Mode)"
    url: "https://{host}:{port}"
    variables:
      host:
        default: "127.0.0.1"
        description: "Binding host"
      port:
        default: "27124"
        description: "HTTPS port"
  - description: "HTTP (Insecure Mode)"
    url: "http://{host}:{port}"
    variables:
      host:
        default: "127.0.0.1"
        description: "Binding host"
      port:
        default: "27123"
        description: "HTTP port"